+++ include/fem-dem/cfd_dem_coupling.h
@@ -15,6 +15,10 @@
  *
  */
 
+
+#ifndef lethe_dem_cfd_coupling_h
+#define lethe_dem_cfd_coupling_h
+
 #include <solvers/navier_stokes_scratch_data.h>
 
 #include <dem/adaptive_sparse_contacts.h>
@@ -39,9 +43,6 @@
 
 #include <deal.II/numerics/vector_tools.h>
 
-#ifndef lethe_dem_cfd_coupling_h
-#  define lethe_dem_cfd_coupling_h
-
 using namespace dealii;
 
 template <int dim>
@@ -49,7 +50,6 @@ class CFDDEMSolver : public GLSVANSSolver<dim>
 {
   using FuncPtrType = bool (CFDDEMSolver<dim>::*)(const unsigned int &counter);
   FuncPtrType check_contact_search_step;
-  FuncPtrType check_load_balance_step;
 
 public:
   CFDDEMSolver(CFDDEMSimulationParameters<dim> &nsparam);
@@ -58,40 +58,6 @@ public:
 
   virtual void
   solve() override;
-
-  /**
-   * The cell_weight() function indicates to the triangulation how much
-   * computational work is expected to happen on this cell, and consequently
-   * how the domain needs to be partitioned so that every MPI rank receives a
-   * roughly equal amount of work (potentially not an equal number of cells).
-   * While the function is called from the outside, it is connected to the
-   * corresponding signal from inside this class, therefore it can be private.
-   * This function is the key component that allow us to dynamically balance the
-   * computational load. The function attributes a weight to
-   * every cell that represents the computational work on this cell. Here the
-   * majority of work is expected to happen on the particles, therefore the
-   * return value of this function (representing "work for this cell") is
-   * calculated based on the number of particles in the current cell.
-   * The function is connected to the cell_weight() signal inside the
-   * triangulation, and will be called once per cell, whenever the triangulation
-   * repartitions the domain between ranks (the connection is created inside the
-   * particles_generation() function of this class).
-   */
-#  if (DEAL_II_VERSION_MAJOR < 10 && DEAL_II_VERSION_MINOR < 6)
-  unsigned int
-  cell_weight(
-    const typename parallel::distributed::Triangulation<dim>::cell_iterator
-                                                                        &cell,
-    const typename parallel::distributed::Triangulation<dim>::CellStatus status)
-    const;
-#  else
-  unsigned int
-  cell_weight(
-    const typename parallel::distributed::Triangulation<dim>::cell_iterator
-                    &cell,
-    const CellStatus status) const;
-#  endif
-
 
   /**
    * @brief Manages the call to the load balancing. Returns true if
@@ -120,12 +86,6 @@ private:
    */
   void
   dem_setup_contact_parameters();
-
-  /**
-   * @brief Connect triangulation for load balancing and particles
-   */
-  void
-  manage_triangulation_connections();
 
   /**
    * @brief Carries out the initialization of DEM parameters
@@ -311,6 +271,7 @@ private:
   double                                     smallest_contact_search_criterion;
 
   DEMContactManager<dim>           contact_manager;
+  LagrangianLoadBalancing<dim>     load_balancing;
   ParticlePointLineForce<dim>      particle_point_line_contact_force_object;
   std::shared_ptr<Integrator<dim>> integrator_object;
   std::shared_ptr<Insertion<dim>>  insertion_object;